@@ -3,14 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
-using Microsoft.AspNet.OData;
-using Microsoft.AspNet.OData.Query;
-using Microsoft.AspNet.OData.Routing;
+using Microsoft.AspNetCore.OData.Query;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Options;
-using Microsoft.OData.Edm;
 using PointOfSalesV2.Api.Security;
 using PointOfSalesV2.Common;
 using PointOfSalesV2.Entities;
@@ -36,59 +33,60 @@ namespace PointOfSalesV2.Api.Controllers
         protected readonly IBase<ExceptionLog> exceptionsRepo;
         protected IEnumerable<LanguageKey> languageKeys;
         protected readonly AppSections section;
-        public BaseController(IOptions<AppSettings> appSettings, IDataRepositoryFactory repositoryFactory,IMemoryCache cache, IBase<T> customRepo = null, AppSections appSections= AppSections.NotSpecified)
+        public BaseController(IOptions<AppSettings> appSettings, IDataRepositoryFactory repositoryFactory, IMemoryCache cache, IBase<T> customRepo = null, AppSections appSections = AppSections.NotSpecified)
         {
             this._cache = cache;
             _appSettings = appSettings;
             _repositoryFactory = repositoryFactory;
             this.languageKeysRepo = repositoryFactory.GetDataRepositories<LanguageKey>();
             this.languageKeys = _cache.Get<IEnumerable<LanguageKey>>("languageKeysMem");
-            if (this.languageKeys == null) 
+            if (this.languageKeys == null)
             {
-                this.languageKeys = this.languageKeysRepo.GetAll(x=>x, y => y.Active == true).Data;
-                this._cache.Set("languageKeysMem", this.languageKeys,DateTime.Now.AddHours(24));
+                this.languageKeys = this.languageKeysRepo.GetAll(x => x, y => y.Active == true).Data;
+                this._cache.Set("languageKeysMem", this.languageKeys, DateTime.Now.AddHours(24));
             }
-            this._baseRepo=customRepo?? _repositoryFactory.GetDataRepositories<T>();
+            this._baseRepo = customRepo ?? _repositoryFactory.GetDataRepositories<T>();
             this.exceptionsRepo = this._repositoryFactory.GetDataRepositories<ExceptionLog>();
             this.section = appSections;
         }
 
         [HttpGet]
         [ActionAuthorize(Operations.READALL)]
-        [EnableQuery()]
+        [EnableQuery]
+        [Microsoft.AspNetCore.OData.Routing.Attributes.ODataAttributeRouting]
         [EnableCors("AllowAllOrigins")]
         public virtual async Task<IActionResult> Get()
         {
             try
             {
-                var data = await _baseRepo.GetAllAsync<T>(x => x.AsNoTracking(),y=>y.Active==true);
+                var data = await _baseRepo.GetAllAsync<T>(x => x.AsNoTracking(), y => y.Active == true);
                 return Ok(data);
             }
 
             catch (Exception ex)
             {
-               await SaveException(ex);
+                await SaveException(ex);
                 return Ok(new { status = -1, message = ex.Message });
             }
         }
 
-        protected async Task SaveException(Exception ex) 
+        protected async Task SaveException(Exception ex)
         {
-          await  this.exceptionsRepo.AddAsync(new ExceptionLog()
+            await this.exceptionsRepo.AddAsync(new ExceptionLog()
             {
                 Code = ex.HResult.ToString(),
                 Active = true,
-                Message=ex.Message.Length>500?ex.Message.Substring(0,499):ex.Message,
-                Section=(int)this.section,
-                SectionName= this.section.ToString()
+                Message = ex.Message.Length > 500 ? ex.Message.Substring(0, 499) : ex.Message,
+                Section = (int)this.section,
+                SectionName = this.section.ToString()
             });
         }
 
         [HttpGet("ExportToExcel")]
         [ActionAuthorize(Operations.EXPORT)]
+        [EnableQuery]
+        [Microsoft.AspNetCore.OData.Routing.Attributes.ODataAttributeRouting]
         [EnableCors("AllowAllOrigins")]
-        [EnableQuery()]
-        [ODataRoute("[controller]/ExportToExcel")]
         public virtual async Task< IActionResult> ExportToExcel()
         {
             try
@@ -121,9 +119,8 @@ namespace PointOfSalesV2.Api.Controllers
 
         [HttpGet("{id:long}")]
         [EnableCors("AllowAllOrigins")]
-        //[EnableQuery]
         [ActionAuthorize(Operations.READ)]
-        public virtual async Task<IActionResult> Get(long id)
+        public virtual async Task<IActionResult> GetById(long id)
         {
             try
             {
@@ -138,26 +135,7 @@ namespace PointOfSalesV2.Api.Controllers
             }
         }
 
-        [HttpGet("{number:int}/{size:int}")]
-        [EnableCors("AllowAllOrigins")]
-        // [EnableQuery]
-        [ActionAuthorize(Operations.READALL)]
-        public virtual async Task<IActionResult> Get(int number, int size)
-        {
-            try
-            {
-                var data = _baseRepo.GetPaged(number, size);
-                data.Status = 0;
-                data.Message = "ok_msg";
-                return Ok(data);
-            }
-
-            catch (Exception ex)
-            {
-              await  SaveException(ex);
-                return Ok(new { status = -1, message = ex.Message });
-            }
-        }
+       
 
 
 
