@@ -10,9 +10,8 @@ using PointOfSalesV2.Entities; using Microsoft.Extensions.Caching.Memory;
 using PointOfSalesV2.Entities.Model;
 using PointOfSalesV2.Repository;
 using static PointOfSalesV2.Common.Enums;
-using Microsoft.AspNet.OData;
-using Microsoft.AspNet.OData.Routing;
 using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.OData.Query;
 
 namespace PointOfSalesV2.Api.Controllers
 {
@@ -37,7 +36,8 @@ namespace PointOfSalesV2.Api.Controllers
         }
 
         [HttpGet]
-        [EnableQuery()]
+        [EnableQuery]
+        [Microsoft.AspNetCore.OData.Routing.Attributes.ODataAttributeRouting]
         [EnableCors("AllowAllOrigins")]
         public  async Task<IActionResult> Get()
         {
@@ -81,7 +81,7 @@ namespace PointOfSalesV2.Api.Controllers
 
         [HttpGet("{id:long}")]
         [EnableCors("AllowAllOrigins")]
-        public  async Task<IActionResult> Get(long id)
+        public   async Task<IActionResult> GetById(long id)
         {
             try
             {
